+++ src/_exports.pyx
@@ -69,9 +69,9 @@ def decode_latin1(object data, object maxdepth=None, object some=False):
     data : bytes
         JSON5 serialized data, encoded as Latin-1 or ASCII.
     maxdepth : Optional[int]
-        see :func:`decode(…) <pyjson5.decode>`
+        see :func:`decode(…) <pyjson5x.decode>`
     some : bool
-        see :func:`decode(…) <pyjson5.decode>`
+        see :func:`decode(…) <pyjson5x.decode>`
 
     Raises
     ------
@@ -83,7 +83,7 @@ def decode_latin1(object data, object maxdepth=None, object some=False):
     Returns
     -------
     object
-        see :func:`decode(…) <pyjson5.decode>`
+        see :func:`decode(…) <pyjson5x.decode>`
     '''
     return decode_buffer(data, maxdepth, bool(some), 1)
 
@@ -101,9 +101,9 @@ def decode_utf8(object data, object maxdepth=None, object some=False):
     data : bytes
         JSON5 serialized data, encoded as UTF-8 or ASCII.
     maxdepth : Optional[int]
-        see :func:`decode(…) <pyjson5.decode>`
+        see :func:`decode(…) <pyjson5x.decode>`
     some : bool
-        see :func:`decode(…) <pyjson5.decode>`
+        see :func:`decode(…) <pyjson5x.decode>`
 
     Raises
     ------
@@ -115,7 +115,7 @@ def decode_utf8(object data, object maxdepth=None, object some=False):
     Returns
     -------
     object
-        see :func:`decode(…) <pyjson5.decode>`
+        see :func:`decode(…) <pyjson5x.decode>`
     '''
     return decode_buffer(data, maxdepth, bool(some), 0)
 
@@ -139,9 +139,9 @@ def decode_buffer(object obj, object maxdepth=None, object some=False,
         The argument must support Python's buffer protocol, i.e.
         :class:`memoryview(…) <memoryview>` must work. The buffer must be contigious.
     maxdepth : Optional[int]
-        see :func:`decode(…) <pyjson5.decode>`
+        see :func:`decode(…) <pyjson5x.decode>`
     some : bool
-        see :func:`decode(…) <pyjson5.decode>`
+        see :func:`decode(…) <pyjson5x.decode>`
     wordlength : Optional[int]
         Must be 0, 1, 2, 4 to denote UTF-8, UCS1, USC2 or USC4 data, resp.
         Surrogates are not supported. Decode the data to an :class:`str` if need be.
@@ -159,7 +159,7 @@ def decode_buffer(object obj, object maxdepth=None, object some=False,
     Returns
     -------
     object
-        see :func:`decode(…) <pyjson5.decode>`
+        see :func:`decode(…) <pyjson5x.decode>`
     '''
     cdef Py_buffer view
 
@@ -201,9 +201,9 @@ def decode_callback(object cb, object maxdepth=None, object some=False,
         * **None:** \
             input exhausted
     maxdepth : Optional[int]
-        see :func:`decode(…) <pyjson5.decode>`
+        see :func:`decode(…) <pyjson5x.decode>`
     some : bool
-        see :func:`decode(…) <pyjson5.decode>`
+        see :func:`decode(…) <pyjson5x.decode>`
     args : Optional[Iterable[Any]]
         Arguments to call ``cb`` with.
 
@@ -217,7 +217,7 @@ def decode_callback(object cb, object maxdepth=None, object some=False,
     Returns
     -------
     object
-        see :func:`decode(…) <pyjson5.decode>`
+        see :func:`decode(…) <pyjson5x.decode>`
     '''
     if not callable(cb):
         raise TypeError(f'type(cb)=={type(cb)!r} is not callable')
@@ -257,9 +257,9 @@ def decode_io(object fp, object maxdepth=None, object some=True):
     fp : IOBase
         A file-like object to parse from.
     maxdepth : Optional[int] = None
-        see :func:`decode(…) <pyjson5.decode>`
+        see :func:`decode(…) <pyjson5x.decode>`
     some : bool
-        see :func:`decode(…) <pyjson5.decode>`
+        see :func:`decode(…) <pyjson5x.decode>`
 
     Raises
     ------
@@ -271,7 +271,7 @@ def decode_io(object fp, object maxdepth=None, object some=True):
     Returns
     -------
     object
-        see :func:`decode(…) <pyjson5.decode>`
+        see :func:`decode(…) <pyjson5x.decode>`
     '''
     if not isinstance(fp, IOBase):
         raise TypeError(f'type(fp)=={type(fp)!r} is not IOBase compatible')
@@ -385,11 +385,11 @@ def encode_bytes(object data, *, options=None, **options_kw):
     Parameters
     ----------
     data : object
-        see :func:`encode(…) <pyjson5.encode>`
+        see :func:`encode(…) <pyjson5x.encode>`
     options : Optional[Options]
-        see :func:`encode(…) <pyjson5.encode>`
+        see :func:`encode(…) <pyjson5x.encode>`
     options_kw
-        see :func:`encode(…) <pyjson5.encode>`
+        see :func:`encode(…) <pyjson5x.encode>`
 
     Raises
     ------
@@ -401,7 +401,7 @@ def encode_bytes(object data, *, options=None, **options_kw):
     Returns
     -------
     bytes
-        see :func:`encode(…) <pyjson5.encode>`
+        see :func:`encode(…) <pyjson5x.encode>`
     '''
     cdef void *temp
     cdef object result
@@ -472,7 +472,7 @@ def encode_callback(object data, object cb, object supply_bytes=False, *,
     Parameters
     ----------
     data : object
-        see :func:`encode(…) <pyjson5.encode>`
+        see :func:`encode(…) <pyjson5x.encode>`
     cb : Callable[[Union[bytes|str]], None]
         A callback function.
         Depending on the truthyness of ``supply_bytes`` either :class:`bytes` or
@@ -481,9 +481,9 @@ def encode_callback(object data, object cb, object supply_bytes=False, *,
         Call ``cb(…)`` with a :class:`bytes` argument if true,
         otherwise :class:`str`.
     options : Optional[Options]
-        see :func:`encode(…) <pyjson5.encode>`
+        see :func:`encode(…) <pyjson5x.encode>`
     options_kw
-        see :func:`encode(…) <pyjson5.encode>`
+        see :func:`encode(…) <pyjson5x.encode>`
 
     Raises
     ------
@@ -521,16 +521,16 @@ def encode_io(object data, object fp, object supply_bytes=True, *,
     Parameters
     ----------
     data : object
-        see :func:`encode(…) <pyjson5.encode>`
+        see :func:`encode(…) <pyjson5x.encode>`
     fp : IOBase
         A file-like object to serialize into.
     supply_bytes : bool
         Call :meth:`fp.write(…) <io.BufferedWriter.write>` with a :class:`bytes` argument if true,
         otherwise :class:`str`.
     options : Optional[Options]
-        see :func:`encode(…) <pyjson5.encode>`
+        see :func:`encode(…) <pyjson5x.encode>`
     options_kw
-        see :func:`encode(…) <pyjson5.encode>`
+        see :func:`encode(…) <pyjson5x.encode>`
 
     Raises
     ------
@@ -579,11 +579,11 @@ def encode_noop(object data, *, options=None, **options_kw):
     Parameters
     ----------
     data : object
-        see :func:`encode(…) <pyjson5.encode>`
+        see :func:`encode(…) <pyjson5x.encode>`
     options : Optional[Options]
-        see :func:`encode(…) <pyjson5.encode>`
+        see :func:`encode(…) <pyjson5x.encode>`
     options_kw
-        see :func:`encode(…) <pyjson5.encode>`
+        see :func:`encode(…) <pyjson5x.encode>`
 
     Returns
     -------