+++ src/_legacy.pyx
@@ -2,7 +2,7 @@ def loads(s, *, encoding='UTF-8', **kw):
     '''
     Decodes JSON5 serialized data from a string.
 
-    Use :func:`decode(…) <pyjson5.decode>` instead!
+    Use :func:`decode(…) <pyjson5x.decode>` instead!
 
     .. code:: python
 
@@ -21,7 +21,7 @@ def loads(s, *, encoding='UTF-8', **kw):
     Returns
     -------
     object
-        see :func:`decode(…) <pyjson5.decode>`
+        see :func:`decode(…) <pyjson5x.decode>`
     '''
     if not isinstance(s, unicode):
         s = PyUnicode_FromEncodedObject(s, encoding, 'strict')
@@ -32,7 +32,7 @@ def load(fp, **kw):
     '''
     Decodes JSON5 serialized data from a file-like object.
 
-    Use :func:`decode_io(…) <pyjson5.decode_io>` instead!
+    Use :func:`decode_io(…) <pyjson5x.decode_io>` instead!
 
     .. code:: python
 
@@ -48,7 +48,7 @@ def load(fp, **kw):
     Returns
     -------
     str
-        see :func:`decode_io(…) <pyjson5.decode_io>`
+        see :func:`decode_io(…) <pyjson5x.decode_io>`
     '''
     return decode_io(fp, None, False)
 
@@ -57,7 +57,7 @@ def dumps(obj, **kw):
     '''
     Serializes a Python object to a JSON5 compatible string.
 
-    Use :func:`encode(…) <pyjson5.encode>` instead!
+    Use :func:`encode(…) <pyjson5x.encode>` instead!
 
     .. code:: python
 
@@ -73,7 +73,7 @@ def dumps(obj, **kw):
     Returns
     -------
     str
-        see :func:`encode(…) <pyjson5.encode>`
+        see :func:`encode(…) <pyjson5x.encode>`
     '''
     return encode(obj)
 
@@ -82,7 +82,7 @@ def dump(object obj, object fp, **kw):
     '''
     Serializes a Python object to a JSON5 compatible string.
 
-    Use :func:`encode_io(…) <pyjson5.encode_io>` instead!
+    Use :func:`encode_io(…) <pyjson5x.encode_io>` instead!
 
     .. code:: python